Refactor Service & Controllers || Reward System
