+++ package.json
@@ -29,6 +29,7 @@
     "multer": "^1.4.5-lts.1",
     "mysql2": "^3.6.1",
     "node-cron": "^3.0.2",
+    "node-schedule": "^2.1.1",
     "nodemon": "^3.0.1",
     "progress": "^2.0.3",
     "rotating-file-stream": "^3.2.1",
