+++ server/src/controllers/rewardController.js
@@ -1,21 +1,22 @@
 const rewardService = require('../services/rewardService');
 const logger = require("../utils/logger/logger");
 
-const allRewards = async function (req, res) {
+const getAllRewards = async function (req, res) {
   try {
     const rewards = rewardService.allRewards();
     res.json(rewards);
-    logger.info("Trophées récupérés:", rewards);
   } catch (error) {
     res.status(500).json({ message: 'Route protégée', error: error.message })
     logger.error('Erreur lors de la recuperation des trophées:', error)
   }
 }
+
 const getAllUserRewards = async function (req, res) {
   const userId = req.params.userId;
   const rewards = await rewardService.getRewardsByUser(userId);
   res.json(rewards);
 }
+
 const checkAvailableRewards = async function (req, res) {
   const userId = req.params.userId;
   const rewards = await rewardService.checkAvailableRewards(userId);
@@ -23,6 +24,6 @@ const checkAvailableRewards = async function (req, res) {
 }
 
 module.exports = {
-  allRewards,
+  getAllRewards,
   getAllUserRewards,
 }
